+++ @geogebra/solver-sdk/src/types.ts
@@ -1,4 +1,4 @@
-import { DecoratorType, NestedExpressionType } from './parser/types';
+import { DecoratorType, ExpressionDecorations, NestedExpressionType } from './parser';
 
 export type PlanId =
   | 'Equations.SolveLinearEquation'
@@ -27,6 +27,12 @@ export type SolverContext = {
   solutionVariable?: string;
 };
 
+export type ApiMathFormat =
+  | 'solver' // solver format - do not use unless you know why you are doing it
+  | 'latex' // latex format - good for quick rendering, but lacks metadata
+  | 'json' // array-based json format, use this one (will be deprecated in favour of json2)
+  | 'json2'; // object-based json format, not supported on all servers yet so use with care
+
 export type API_VERSION_INFO_RESPONSE = {
   commit: string;
   deploymentName: string;
@@ -45,7 +51,12 @@ type PlanSelectionBase<MathFormat> = {
 export type PlanSelectionSolver = PlanSelectionBase<string>;
 export type PlanSelectionLatex = PlanSelectionBase<string>;
 export type PlanSelectionJson = PlanSelectionBase<MathJson>;
-export type PlanSelection = PlanSelectionSolver | PlanSelectionLatex | PlanSelectionJson;
+export type PlanSelectionJson2 = PlanSelectionBase<MathJson2>;
+export type PlanSelection =
+  | PlanSelectionSolver
+  | PlanSelectionLatex
+  | PlanSelectionJson
+  | PlanSelectionJson2;
 
 export type API_SELECT_PLANS_RESPONSE = PlanSelection[];
 
@@ -64,6 +75,7 @@ export type PathMapping = {
   toPaths: string[];
 };
 
+// This is how the `json` format is typed.
 export type MathJson =
   | [string] // variable, number, undefined, infinity, or reals
   | [[string, ...(DecoratorType | string)[]]] // variable or number
@@ -73,6 +85,25 @@ export type MathJson =
   | [['SmartProduct', ...(DecoratorType | string)[]], ...SmartProductOperandJson[]];
 
 export type SmartProductOperandJson = [boolean, MathJson];
+
+// This is how the `json2` format is typed. This For a human-centric description of this format,
+// see docs/expression-serialization.md
+export type MathJson2 = ExpressionDecorations &
+  (
+    | {
+        type: NestedExpressionType;
+        operands?: MathJson2[];
+      }
+    | {
+        type: 'Integer' | 'Decimal' | 'RecurringDecimal' | 'Variable' | 'Name';
+        value: string;
+      }
+    | {
+        type: 'SmartProduct';
+        operands: MathJson2[];
+        signs: boolean[];
+      }
+  );
 
 export type GmAction = {
   type:
@@ -126,17 +157,22 @@ export type TaskBase<MathFormat> = {
 export type TransformationSolver = TransformationBase<string>;
 export type TransformationLatex = TransformationBase<string>;
 export type TransformationJson = TransformationBase<MathJson>;
+export type TransformationJson2 = TransformationBase<MathJson2>;
 export type Transformation =
   | TransformationSolver
   | TransformationLatex
-  | TransformationJson;
+  | TransformationJson
+  | TransformationJson2;
 
 export type TaskSolver = TaskBase<string>;
 export type TaskLatex = TaskBase<string>;
 export type TaskJson = TaskBase<MathJson>;
-export type Task = TaskSolver | TaskLatex | TaskJson;
+export type TaskJson2 = TaskBase<MathJson2>;
+export type Task = TaskSolver | TaskLatex | TaskJson | TaskJson2;
 
 export type MetadataSolver = MetadataBase<string>;
 export type MetadataLatex = MetadataBase<string>;
 export type MetadataJson = MetadataBase<MathJson>;
-export type Metadata = MetadataSolver | MetadataLatex | MetadataJson;
+export type MetadataJson2 = MetadataBase<MathJson2>;
+
+export type Metadata = MetadataSolver | MetadataLatex | MetadataJson | MetadataJson2;